refactor: move getPostRelativeDir to post-utils and clean up url-utils
