+++ src/pages/posts/[...slug].astro
@@ -12,7 +12,8 @@ import MainGridLayout from "@layouts/MainGridLayout.astro";
 
 import { getSortedPosts } from "@utils/content-utils";
 
-import { getPostRelativeDir, getPostUrlById } from "@utils/url-utils";
+import { getPostRelativeDir } from "@utils/post-utils";
+import { getPostUrlById } from "@utils/url-utils";
 
 import { Icon } from "astro-icon/components";
